refactor: remove redundant component prefix
