+++ frontend/src/components/app/Banner.vue
