+++ frontend/src/components/app/Footer.vue
