+++ frontend/src/components/app/Logo.vue
