+++ frontend/src/components/app/SkipLinks.vue
