+++ frontend/src/layouts/base.vue
@@ -1,8 +1,6 @@
 <script setup lang="ts">
 import { NuxtLoadingIndicator } from "#components";
 import ConsentBanner from "~/components/analytics/ConsentBanner.vue";
-import AppFooter from "~/components/app/AppFooter.vue";
-import AppHeader from "~/components/app/AppHeader.vue";
 </script>
 
 <template>
